requirements.txt
requests
beautifulsoup4
pdfplumber
discord.py